Additional dispatcher removal simplifications and waiting repear updates
